Data type fix for timestamp data

+++ Squad.Bot/Models/AI/MemberVoiceActivity.cs
@@ -16,7 +16,7 @@ namespace Squad.Bot.Models.AI
         [Required]
         public Guilds Guilds { get; set; } = null!;
 
-        public int TotalMinutes { get; set; } = 0;
+        public double TotalMinutes { get; set; } = 0;
 
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     }
